削除の条件分岐

+++ app/controllers/items_controller.rb
@@ -36,7 +36,9 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    redirect_to root_path if item.destroy
+    if item.user_id == current_user.id
+      redirect_to root_path if item.destroy
+    end
   end
 
   private
